# specs/002-github-app/contracts/api.openapi.yaml
openapi: 3.0.3
info:
  title: RepoWeaver GitHub App API
  description: REST API for RepoWeaver GitHub App web interface and webhook handling
  version: 1.0.0
  contact:
    name: RepoWeaver
    url: https://github.com/pradeepmouli/repoweaver
servers:
  - url: https://app.repoweaver.dev
    description: Production
  - url: http://localhost:3000
    description: Local development
security:
  - sessionToken: []
  - githubWebhookSignature: []
paths:
...
  /auth/github/callback:
    get:
      summary: GitHub OAuth callback
      tags: [Authentication]
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        '302':
          description: Redirect to dashboard with session cookie
          headers:
            Set-Cookie:
              schema:
                type: string
                example: session=abc123; HttpOnly; Secure
        '401':
          description: OAuth authorization failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      summary: Logout and invalidate session
      tags: [Authentication]
      responses:
        '200':
          description: Session invalidated
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================================================
  # Installations & Repositories
  # ============================================================================
...
components:
  securitySchemes:
    sessionToken:
      type: apiKey
      in: cookie
      name: session
    githubWebhookSignature:
      type: apiKey
      in: header
      name: X-Hub-Signature-256
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...
  responses:
    Unauthorized:
      description: Unauthorized - invalid or missing session
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Please log in to continue
...